[FIX] exported PubAck, isNatsError, checkJsError [FIX] JetStream `pull()` tests incorrectly set expires as `nanos(expire)` in `PullOptions` - values on the client are expressed as millis, and converted internally. Examples showed correctly values as millis [update] nats-server to 2.3.3 [breaking] internal kv implementation changes `keys():Promise<Set<string>>` tp `keys():Promise<string[]>`, this is non-public api.
