+++ .github/workflows/test.yml
@@ -30,7 +30,7 @@ jobs:
           deno-version: ${{ matrix.deno-version }}
 
       - name: Set NATS Server Version
-        run: echo "NATS_VERSION=v2.3.2" >> $GITHUB_ENV
+        run: echo "NATS_VERSION=v2.3.3" >> $GITHUB_ENV
 
       - name: Get nats-server
         run: |
