+++ nats-base-client/mod.ts
@@ -3,6 +3,7 @@ export {
   AdvisoryKind,
   Bench,
   canonicalMIMEHeaderKey,
+  checkJsError,
   consumerOpts,
   createInbox,
   credsAuthenticator,
@@ -66,6 +67,7 @@ export type {
   NatsConnection,
   PeerInfo,
   Placement,
+  PubAck,
   PublishOptions,
   PullOptions,
   RequestOptions,
